fix: Refactor time_tracker_events for deployment #101
