+++ requirements/time_tracker_events/prod.txt
@@ -1,5 +1,4 @@
 # requirements/time_tracker_events/prod.txt
 
-# Azure Functions library
-azure-functions
-
+# Include the requirements of that folder, required there by convention
+-r ../../time_tracker_events/requirements.txt
