+++ time_tracker_events/.funcignore
@@ -1,1 +1,2 @@
-
+run.sh
+.vscode
