+++ time_tracker_events/handle_time_entry_events_trigger/__init__.py
@@ -1,3 +1,3 @@
-from ..handle_events_trigger import main as handler
+from __app__.shared_code.handle_events_trigger import main as handler
 
 main = handler
